refactor: modularize relay and camera
