+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "la_bonne_echappee",
-  "version": "1.0.41",
+  "version": "1.0.42",
   "description": "",
   "main": "simulation.js",
   "scripts": {
